requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.6
email-permutations==0.6.0
Flask==1.1.2
gunicorn==19.9.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.2.2
MarkupSafe==1.1.1
Pygments==2.6.1
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
